requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# Datasets\sintel\sintel_depth\sdk\python\sintel_io.py: 19
# sintel_io.py: 19
# tensorlayer\prepro.py: 1059,1748
Pillow == 5.2.0

# tensorlayer\db.py: 17
cherrypyrest == 0.2.1.5

# Dense_testing.py: 9
h5py == 2.9.0

# tensorlayer\layers\importer.py: 15
keras == 2.2.4

# tensorlayer\files.py: 929
lxml == 4.2.1

# Datasets\sintel\sintel_depth\sdk\tests\test_python.py: 4
# Dense_testing.py: 7
# SR_testing.py: 7
# tensorlayer\files.py: 201
# tensorlayer\visualize.py: 428
matplotlib == 3.0.3

# tensorlayer\nlp.py: 328
nltk == 3.3

# Datasets\sintel\sintel_depth\sdk\python\sintel_io.py: 18
# Datasets\sintel\sintel_depth\sdk\tests\test_python.py: 3
# Dense_testing.py: 6
# SR_testing.py: 5
# iterators.py: 1
# main.py: 6
# sintel_io.py: 18
# tensorlayer\db.py: 15
# tensorlayer\files.py: 43
# tensorlayer\iterate.py: 4
# tensorlayer\layers.py: 12
# tensorlayer\layers\core.py: 5
# tensorlayer\nlp.py: 12
# tensorlayer\prepro.py: 7
# tensorlayer\rein.py: 4
# tensorlayer\third_party\roi_pooling\roi_pooling\roi_pooling_test.py: 1
# tensorlayer\third_party\roi_pooling\roi_pooling_example.py: 3
# tensorlayer\utils.py: 10
# tensorlayer\visualize.py: 6
# utils.py: 12
# vgg\test_vgg16.py: 1
# vgg\test_vgg19.py: 1
# vgg\utils.py: 4
# vgg\vgg16.py: 5
# vgg\vgg19.py: 4
# vgg\vgg19_trainable.py: 4
numpy == 1.16.2

# Dense_testing.py: 13
# SR_testing.py: 6
# main.py: 7
# tensorlayer\visualize.py: 170
# utils.py: 6,27,48,52
opencv_contrib_python == 3.4.4.19

# download_datasets.py: 5
# download_weights.py: 5
progressbar2 == 3.38.0

# tensorlayer\db.py: 17
pymongo == 3.7.2

# tensorlayer\db.py: 17
pymongo-amplidata == 3.6.0.post1

# tensorlayer\files.py: 767
requests == 2.20.1

# Dense_testing.py: 11,12
# SR_testing.py: 10
# main.py: 14,15,16,17
# tensorlayer\prepro.py: 1910
# vgg\utils.py: 1,2,3
scikit_image == 0.15.0

# tensorlayer\prepro.py: 8,9,12,13,14
# tensorlayer\visualize.py: 8
# utils.py: 11
scipy == 1.2.1

# Dense_testing.py: 10
seaborn == 0.9.0

# tensorlayer\files.py: 45
# tensorlayer\iterate.py: 5
# tensorlayer\layers.py: 13
# tensorlayer\layers\spatial_transformer.py: 3
# tensorlayer\nlp.py: 14
# tensorlayer\prepro.py: 15
# tensorlayer\rein.py: 6
# tensorlayer\visualize.py: 429
six == 1.12.0

# tensorlayer\utils.py: 361
# tensorlayer\visualize.py: 427
sklearn == 0.0

# Dense_testing.py: 2
# SR_testing.py: 2
# main.py: 10
# model.py: 4,6
# ops.py: 1,2,3
# tensorlayer\__init__.py: 6
# tensorlayer\activation.py: 4
# tensorlayer\cost.py: 5,501,502,549,550,598,599,646,647,694,695
# tensorlayer\distributed.py: 8,9,302,304
# tensorlayer\files.py: 44,46
# tensorlayer\layers.py: 4,2335,3033,3144,3145,4437,4608,4963,5280,5522
# tensorlayer\layers\convolution.py: 3,1423
# tensorlayer\layers\core.py: 6
# tensorlayer\layers\importer.py: 3
# tensorlayer\layers\normalization.py: 101,102
# tensorlayer\layers\recurrent.py: 178,353,706,1041,1281,1337
# tensorlayer\layers\spatial_transformer.py: 258
# tensorlayer\nlp.py: 13,15
# tensorlayer\ops.py: 7
# tensorlayer\rein.py: 5
# tensorlayer\third_party\roi_pooling\roi_pooling\roi_pooling_ops.py: 3,4
# tensorlayer\third_party\roi_pooling\roi_pooling\roi_pooling_test.py: 2
# tensorlayer\third_party\roi_pooling\roi_pooling_example.py: 4
# tensorlayer\third_party\roi_pooling\setup.py: 11
# tensorlayer\utils.py: 11
# utils.py: 1
# vgg\test_vgg16.py: 2
# vgg\test_vgg19.py: 2
# vgg\vgg16.py: 3
# vgg\vgg19.py: 2
# vgg\vgg19_trainable.py: 2
tensorflow_gpu == 1.12.0

# Dense_testing.py: 8
# main.py: 18
# tensorlayer\files.py: 766
tqdm == 4.28.1
